Create ShopSessionController for the shop

+++ app/Http/Controllers/API/ProductController.php
@@ -6,7 +6,6 @@ use Exception;
 use App\Models\Image;
 use App\Models\Product;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\DB;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Auth;
 
@@ -44,6 +43,7 @@ class ProductController extends Controller
     {
         try {
             if (Auth::user()->role_id === 2) {
+
                 $request->validate([
                     'product_name' => 'required|string|min:3|max:50|unique:products',
                     'product_content' => 'required',
@@ -51,12 +51,26 @@ class ProductController extends Controller
                     'quantity' => 'required|integer',
                     'category_id' => 'required',
                 ]);
-                $product = Product::create($request->all());
+
+                $stripe = new \Stripe\StripeClient(env('STRIPE_SECRET'));
+                $stripes = $stripe->products->create([
+                    'name' => $request->product_name,
+                    'description' => $request->product_content,
+                    'tax_code' => 'txcd_35010000',
+                    'default_price_data' =>
+                    [
+                        'currency' => 'EUR',
+                        'unit_amount' => $request->price * 100,
+                    ],
+                ]);
+
+                $product = Product::create(array_merge($request->all(), ['stripe_id' => $stripes->id]));
 
                 return response()->json([
                     'code' => 201,
                     'status' => 'success',
                     'data' => $product,
+                    'stripe' => $stripes,
                     'message' => 'Ajout du produit avec succès '
                 ]);
             } else {
@@ -147,12 +161,23 @@ class ProductController extends Controller
     public function destroy(Product $product)
     {
         try {
-            $product->delete();
-            return response()->json([
-                'code' => 201,
-                'status' => 'success',
-                'message' => 'Suppression réussite'
-            ]);
+            if (Auth::user()->role_id === 2) {
+                $stripe = new \Stripe\StripeClient(env('STRIPE_SECRET'));
+                $stripe->products->update($product->stripe_id, ['active' => false]);
+                $product->delete();
+
+                return response()->json([
+                    'code' => 201,
+                    'status' => 'success',
+                    'message' => 'Suppression réussite'
+                ]);
+            } else {
+                return response()->json([
+                    'code' => 401,
+                    'status' => 'error',
+                    'message' => 'Pas autorisé',
+                ]);
+            }
         } catch (Exception $e) {
             return response()->json([
                 'code' => 404,
